Допилил GetList

+++ pkg/sql/store/repository.go
@@ -9,7 +9,7 @@ type UserRepository interface {
 	Find(int64) (*model.User, error)
 	Delete(int64) error
 	UpdateUser(*model.User) error
-	GetUserList(page int, limit int) ([]*model.User, error)
+	GetUserList(page int64, limit int64) ([]*model.User, error)
 	SearchPermissionsForUsers(users []*model.User) error
 }
 
@@ -17,7 +17,7 @@ type RoleRepository interface {
 	FindRoleById(int64) (*model.Role, error)
 	CreateRole(string) (*model.Role, error)     // CreateRole (setter)
 	FindRoleByName(string) (*model.Role, error) // FindRoleByName -> Set RolePermission
-	GetRoleList(page int, limit int) ([]*model.Role, error)
+	GetRoleList(page int64, limit int64) ([]*model.Role, error)
 	SearchPermissionsForRoles(roles []*model.Role) error
 }
